corrected type conversion
and fixed unecessary cannot stat warning in installer

+++ install-chunkymap-on-ubuntu.sh
@@ -1,10 +1,12 @@
 #!/bin/sh
 cd $HOME
 MT_MY_WEBSITE_PATH=/var/www/html/minetest
-rm -Rf $HOME/minetest-stuff/minetest-chunkymap
+if [ -d "$HOME/minetest-stuff/minetest-chunkymap" ]; then
+  rm -Rf $HOME/minetest-stuff/minetest-chunkymap
+fi
 CHUNKYMAP_INSTALLER_PATH=$HOME/Downloads/minetest-chunkymap
 if [ ! -d "$HOME/Downloads" ]; then
-	mkdir "$HOME/Downloads"
+  mkdir "$HOME/Downloads"
 fi
 
 MINETEST_UTIL=$HOME/minetest/util
